add validations , security and image managements

+++ src/main/java/ma/youcode/marsoul/controller/SocietyController.java
@@ -2,7 +2,7 @@ package ma.youcode.marsoul.controller;
 
 import ma.youcode.marsoul.dto.SocietyDTO;
 import ma.youcode.marsoul.entity.Society;
-import ma.youcode.marsoul.service.impl.SocietyService;
+import ma.youcode.marsoul.service.SocietyService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -29,7 +29,7 @@ public class SocietyController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<SocietyDTO> getSociety(@PathVariable("id") Integer id) {
+    public ResponseEntity<SocietyDTO> getSociety(@PathVariable("id") Long id) {
         Society society = societyService.getSocietyById(id);
         return ResponseEntity.ok().body(modelMapper.map(society, SocietyDTO.class));
     }
@@ -41,13 +41,13 @@ public class SocietyController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Society> updateSociety(@PathVariable("id") Integer id, @RequestBody SocietyDTO societyDTO) {
+    public ResponseEntity<Society> updateSociety(@PathVariable("id") Long id, @RequestBody SocietyDTO societyDTO) {
         Society society = societyService.updateSociety(id, modelMapper.map(societyDTO, Society.class));
         return ResponseEntity.status(HttpStatus.ACCEPTED).body(society);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<HttpStatus> deleteSociety(@PathVariable("id") Integer id) {
+    public ResponseEntity<HttpStatus> deleteSociety(@PathVariable("id") Long id) {
         societyService.deleteSocietyById(id);
         return ResponseEntity.noContent().build();
     }
